+++ libraries/ghc-boot-th/changelog.md
@@ -1,3 +1,9 @@
+## next *TBA*
+
+  * Add `GHC.OnOff.Type` module
+
+  * Add `Data` and `Ord` instances for `Extension`
+
 ## 8.0.1  *May 2016*
 
   * Bundled with GHC 8.0.1