+++ libraries/ghc-boot/changelog.md
@@ -1,3 +1,7 @@
+## next *TBA*
+
+  * Add `GHC.OnOff` module
+
 ## 8.0.1  *May 2016*
 
   * Bundled with GHC 8.0.1